sdk(elixir): Move download engine to a new module
Reducing provision complexity in `Dagger.EngineConn` into
`Dagger.Internal.Engine.Downloader` and also allow download to
custom cli host for future testing.
